requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
accelerate>=0.25.0
bitsandbytes>=0.42.0

# Image processing
pillow>=9.0.0
requests>=2.25.0

# System utilities
psutil>=5.8.0
gputil>=1.4.0

# Text processing
sentencepiece>=0.1.97
protobuf>=3.20.0

# Optional for better performance
optimum>=1.16.0
